PubKeyResponse made into PublicKey
