+++ tendermint/src/public_key.rs
@@ -1,6 +1,7 @@
 //! Public keys used in Tendermint networks
 
 pub use ed25519_dalek::PublicKey as Ed25519;
+mod pub_key_request;
 
 #[cfg(feature = "secp256k1")]
 pub use k256::EncodedPoint as Secp256k1;
@@ -12,8 +13,18 @@ use crate::{
 use anomaly::{fail, format_err};
 use serde::{de, ser, Deserialize, Serialize};
 use signature::Verifier as _;
+use std::convert::TryFrom;
 use std::{cmp::Ordering, fmt, ops::Deref, str::FromStr};
 use subtle_encoding::{base64, bech32, hex};
+use tendermint_proto::crypto::public_key::Sum;
+use tendermint_proto::privval::PubKeyResponse as RawPubKeyResponse;
+use tendermint_proto::DomainType;
+
+// Note:On the golang side this is generic in the sense that it could everything that implements
+// github.com/tendermint/tendermint/crypto.PubKey
+// While this is meant to be used with different key-types, it currently only uses a PubKeyEd25519
+// version.
+// TODO(ismail): make this more generic (by modifying prost and adding a trait for PubKey)
 
 /// Public keys allowed in Tendermint protocols
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
@@ -37,6 +48,44 @@ pub enum PublicKey {
         deserialize_with = "deserialize_secp256k1_base64"
     )]
     Secp256k1(Secp256k1),
+}
+
+impl DomainType<RawPubKeyResponse> for PublicKey {}
+
+impl TryFrom<RawPubKeyResponse> for PublicKey {
+    type Error = Error;
+
+    // This does not check if the underlying pub_key_ed25519 has the right size.
+    // The caller needs to make sure that this is actually the case.
+    fn try_from(value: RawPubKeyResponse) -> Result<Self, Self::Error> {
+        match &value
+            .pub_key
+            .ok_or_else(|| format_err!(error::Kind::InvalidKey, "empty pubkey"))?
+            .sum
+            .ok_or_else(|| format_err!(error::Kind::InvalidKey, "empty sum"))?
+        {
+            Sum::Ed25519(b) => Ed25519::from_bytes(b),
+        }
+        .map(Into::into)
+        .map_err(|_| format_err!(error::Kind::InvalidKey, "malformed key").into())
+    }
+}
+
+impl From<PublicKey> for RawPubKeyResponse {
+    fn from(value: PublicKey) -> Self {
+        match value {
+            PublicKey::Ed25519(ref pk) => RawPubKeyResponse {
+                pub_key: Some(tendermint_proto::crypto::PublicKey {
+                    sum: Some(tendermint_proto::crypto::public_key::Sum::Ed25519(
+                        pk.as_bytes().to_vec(),
+                    )),
+                }),
+                error: None,
+            },
+            #[cfg(feature = "secp256k1")]
+            PublicKey::Secp256k1(_) => panic!("secp256k1 PubKeyResponse unimplemented"),
+        }
+    }
 }
 
 impl PublicKey {
@@ -318,7 +367,10 @@ where
 #[cfg(test)]
 mod tests {
     use super::{PublicKey, TendermintKey};
+    use crate::public_key::PublicKey::Ed25519;
+    pub use ed25519_dalek::PublicKey as Ed25519PublicKey;
     use subtle_encoding::hex;
+    use tendermint_proto::DomainType;
 
     const EXAMPLE_CONSENSUS_KEY: &str =
         "4A25C6640A1F72B9C975338294EF51B6D1C33158BB6ECBA69FBC3FB5A33C9DCE";
@@ -367,5 +419,50 @@ mod tests {
 
         let reserialized_json = serde_json::to_string(&pubkey).unwrap();
         assert_eq!(reserialized_json.as_str(), json_string);
+    }
+
+    #[test]
+    fn test_ed25519_pubkey_msg() {
+        // test-vector generated from Go
+        /*
+           import (
+               "fmt"
+               "github.com/tendermint/tendermint/proto/tendermint/crypto"
+               "github.com/tendermint/tendermint/proto/tendermint/privval"
+           )
+
+            func ed25519_key() {
+                pkr := &privval.PubKeyResponse{
+                    PubKey: &crypto.PublicKey{
+                        Sum: &crypto.PublicKey_Ed25519{Ed25519: []byte{
+                            215, 90, 152, 1, 130, 177, 10, 183, 213, 75, 254, 211, 201, 100, 7, 58,
+                            14, 225, 114, 243, 218, 166, 35, 37, 175, 2, 26, 104, 247, 7, 81, 26,
+                        },
+                        },
+                    },
+                    Error: nil,
+                }
+                pbpk, _ := pkr.Marshal()
+                fmt.Printf("%#v\n", pbpk)
+
+            }
+        */
+        let encoded = vec![
+            0xa, 0x22, 0xa, 0x20, 0xd7, 0x5a, 0x98, 0x1, 0x82, 0xb1, 0xa, 0xb7, 0xd5, 0x4b, 0xfe,
+            0xd3, 0xc9, 0x64, 0x7, 0x3a, 0xe, 0xe1, 0x72, 0xf3, 0xda, 0xa6, 0x23, 0x25, 0xaf, 0x2,
+            0x1a, 0x68, 0xf7, 0x7, 0x51, 0x1a,
+        ];
+
+        let msg = Ed25519(
+            Ed25519PublicKey::from_bytes(&[
+                215, 90, 152, 1, 130, 177, 10, 183, 213, 75, 254, 211, 201, 100, 7, 58, 14, 225,
+                114, 243, 218, 166, 35, 37, 175, 2, 26, 104, 247, 7, 81, 26,
+            ])
+            .unwrap(),
+        );
+        let got = msg.encode_vec().unwrap();
+
+        assert_eq!(got, encoded);
+        assert_eq!(PublicKey::decode_vec(&encoded).unwrap(), msg);
     }
 }
